requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.7
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Pillow==10.2.0
markdown2==2.4.10
pillow-heif==0.15.0
cryptography==42.0.2
python-dateutil==2.8.2 